requirements.txt
fastapi==0.95.2
pydantic
supabase==1.0.3
langchain==0.0.166
openai==0.27.6
uvicorn==0.22.0
python-dotenv
tiktoken